Update bootstrap-strength-meter.js
This update prevents overload of the browser when typing long passwords. I applied the update only to `progressBar`.  This prevents the password from being scored until a timeout of 500ms has elapsed. Was causing major lag in browser/mobile when typing in a large password quickly.

The `text` method probably still needs to be updated.